+++ src/main/java/site/courseregistrationsystem/auth/presentation/AuthController.java
@@ -1,6 +1,10 @@
 package site.courseregistrationsystem.auth.presentation;
 
+import static site.courseregistrationsystem.util.ProjectConstant.*;
+
 import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -12,6 +16,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import site.courseregistrationsystem.auth.StudentSession;
 import site.courseregistrationsystem.auth.application.AuthService;
+import site.courseregistrationsystem.auth.application.SessionManager;
 import site.courseregistrationsystem.auth.dto.LoginForm;
 import site.courseregistrationsystem.util.api.ApiResponse;
 import site.courseregistrationsystem.util.api.ResponseMessage;
@@ -22,6 +27,7 @@ public class AuthController {
 
 	private final CookieProperties cookieProperties;
 	private final AuthService authService;
+	private final SessionManager sessionManager;
 
 	@ResponseStatus(HttpStatus.CREATED)
 	@PostMapping("/login")
@@ -38,6 +44,37 @@ public class AuthController {
 		response.addCookie(cookie); // 쿠키 설정
 
 		return ApiResponse.of(HttpStatus.CREATED, ResponseMessage.LOGIN_SUCCESS.getMessage(), null);
+	}
+
+	@ResponseStatus(HttpStatus.CREATED)
+	@PostMapping("/session")
+	public ApiResponse<Void> renewSessionDuration(@CookieValue(value = SESSION_ID) Cookie sessionCookie, HttpServletResponse response) {
+		StudentSession renewSession = sessionManager.renew(sessionCookie.getValue());
+
+		Cookie cookie = new Cookie(cookieProperties.getName(), renewSession.getId());
+		cookie.setHttpOnly(true);
+		cookie.setSecure(true);
+		cookie.setDomain(cookieProperties.getDomain());
+		cookie.setPath(cookieProperties.getPath());
+		cookie.setMaxAge(cookieProperties.getExpiry());
+
+		response.addCookie(cookie); // 쿠키 설정
+
+		return ApiResponse.of(HttpStatus.CREATED, ResponseMessage.RENEW_SESSION_DURATION_SUCCESS.getMessage(), null);
+	}
+
+	@DeleteMapping("/logout")
+	public ApiResponse<Void> logout(@CookieValue(value = SESSION_ID) Cookie sessionCookie, HttpServletResponse response) {
+		sessionManager.invalidate(sessionCookie.getValue());
+
+		sessionCookie.setMaxAge(0);
+		sessionCookie.setHttpOnly(true);
+		sessionCookie.setSecure(true);
+		sessionCookie.setDomain(cookieProperties.getDomain());
+		sessionCookie.setPath(cookieProperties.getPath());
+		response.addCookie(sessionCookie);
+
+		return ApiResponse.ok(ResponseMessage.LOGOUT_SUCCESS.getMessage(), null);
 	}
 
 }
