+++ src/test/java/site/courseregistrationsystem/RestDocsSupport.java
@@ -16,7 +16,6 @@ import site.courseregistrationsystem.auth.application.AuthService;
 import site.courseregistrationsystem.auth.application.SessionManager;
 import site.courseregistrationsystem.auth.presentation.AuthController;
 import site.courseregistrationsystem.auth.presentation.CookieProperties;
-import site.courseregistrationsystem.auth.presentation.SessionController;
 import site.courseregistrationsystem.clock.presentation.ClockController;
 import site.courseregistrationsystem.exception.auth.NonexistenceSessionException;
 import site.courseregistrationsystem.student.application.StudentService;
@@ -26,8 +25,7 @@ import site.courseregistrationsystem.util.encryption.Aes256Manager;
 @WebMvcTest(controllers = {
 	AuthController.class,
 	ClockController.class,
-	StudentController.class,
-	SessionController.class
+	StudentController.class
 })
 @AutoConfigureRestDocs
 public abstract class RestDocsSupport {
