--- src/test/java/site/courseregistrationsystem/auth/presentation/SessionControllerTest.java
@@ -1,81 +1,0 @@
-package site.courseregistrationsystem.auth.presentation;
-
-import static org.mockito.BDDMockito.*;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.*;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.restdocs.payload.JsonFieldType;
-
-import jakarta.servlet.http.Cookie;
-import site.courseregistrationsystem.RestDocsSupport;
-import site.courseregistrationsystem.auth.StudentSession;
-
-class SessionControllerTest extends RestDocsSupport {
-
-	@Test
-	@DisplayName("세션 지속 시간 갱신 : 성공")
-	void renewSessionDuration() throws Exception {
-		// given
-		String COOKIE_NAME = "SESSIONID";
-		String COOKIE_VALUE = "03166dc4-2c82-4e55-85f5-f47919f367a6";
-		Cookie sessionCookie = new Cookie(COOKIE_NAME, COOKIE_VALUE);
-
-		String RENEW_UUID = "12346dc4-2c82-4e55-1234-123419f367a6";
-		Long STUDENT_PK = 1L;
-		Long EXPIRATION = 3600L;
-
-		String cookieName = cookieProperties.getName();
-
-		given(sessionManager.renew(anyString()))
-			.willReturn(new StudentSession(RENEW_UUID, STUDENT_PK, EXPIRATION));
-
-		// when & then
-		mockMvc.perform(post("/session")
-				.cookie(sessionCookie))
-			.andDo(print())
-			.andExpect(status().isCreated())
-			.andExpect(cookie().exists(cookieName))
-			.andExpect(cookie().httpOnly(cookieName, true))
-			.andExpect(cookie().secure(cookieName, true))
-			.andExpect(cookie().domain(cookieName, cookieProperties.getDomain()))
-			.andExpect(cookie().path(cookieName, cookieProperties.getPath()))
-			.andExpect(cookie().maxAge(cookieName, cookieProperties.getExpiry()))
-			.andExpect(cookie().value(cookieName, RENEW_UUID))
-			.andDo(document("session-duration-renew-success",
-				preprocessRequest(prettyPrint()),
-				preprocessResponse(prettyPrint()),
-				responseFields(
-					fieldWithPath("code").type(JsonFieldType.NUMBER).description("코드"),
-					fieldWithPath("status").type(JsonFieldType.STRING).description("상태"),
-					fieldWithPath("message").type(JsonFieldType.STRING).description("메시지"),
-					fieldWithPath("data").type(JsonFieldType.NULL).description("응답 데이터")
-				)
-			));
-	}
-
-	@Test
-	@DisplayName("세션 지속 시간 갱신 : 실패")
-	void failRenewSessionDuration() throws Exception {
-		// when & then
-		mockMvc.perform(post("/session"))
-			.andDo(print())
-			.andExpect(status().isUnauthorized())
-			.andDo(document("session-duration-renew-fail",
-				preprocessRequest(prettyPrint()),
-				preprocessResponse(prettyPrint()),
-				responseFields(
-					fieldWithPath("code").type(JsonFieldType.NUMBER).description("코드"),
-					fieldWithPath("status").type(JsonFieldType.STRING).description("상태"),
-					fieldWithPath("message").type(JsonFieldType.STRING).description("메시지"),
-					fieldWithPath("data").type(JsonFieldType.NULL).description("응답 데이터")
-				)
-			));
-	}
-
-}
